+++ confu/schema/__init__.py
@@ -1,2 +1,2 @@
 from confu.schema.core import *
-from confu.schema.inet import Email, Url
+from confu.schema.inet import Email, Url, IpAddress